ml-service/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
joblib
scikit-learn
pydantic
xgboost
numpy<2.0
lightgbm
optuna
